feature: fetch data on component mount for Banner, Categories, Partners, and Promocode management
