+++ client/src/components/admin/CategoriesManagement/CategoriesManagement.jsx
@@ -1,12 +1,20 @@
 import Button from '@components/UI/Button/Button';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import cl from './index.module.scss';
 import PurposeList from './Purpose/PurposeList/PurposeList';
 import TypeList from './Type/TypeList/TypeList';
+import { useDispatch } from 'react-redux';
+import { fetchCategories, fetchTypes } from '@redux/params/service';
 
 const CategoriesManagement = () => {
   const [currentCategory, setCurrentCategory] = useState('purpose');
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchTypes());
+    dispatch(fetchCategories());
+  }, [dispatch]);
 
   return (
     <div className={cl.categoriesManagement}>
