+++ client/src/components/admin/PartnersManagement.jsx/PartnersManagement.jsx
@@ -2,8 +2,17 @@ import { Link } from 'react-router-dom';
 import ReturnButton from '@components/admin/ReturnButton/ReturnButton';
 import List from './PartnersList/PartnersList';
 import cl from './index.module.scss';
+import { useDispatch } from 'react-redux';
+import { useEffect } from 'react';
+import { fetchPartners } from '@redux/partners/service';
 
 const PartnersManagement = () => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchPartners());
+  }, [dispatch]);
+
   return (
     <div className={cl.partnersManagement}>
       <div>
