+++ client/src/components/admin/PromocodeManagement/PromocodeManagement.jsx
@@ -2,11 +2,19 @@ import { Link } from 'react-router-dom';
 import cl from './index.module.scss';
 import List from './PromocodeList/PromocodeList.jsx';
 import PromocodeSelect from './PromocodeSelect/PromocodeSelect';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { responsePromocodeList } from '@redux/admin/selectors';
+import { useEffect } from 'react';
+import { fetchPromocode } from '@redux/promocode/service';
 
 const PromocodeManagement = () => {
   const items = useSelector(responsePromocodeList);
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchPromocode());
+  }, [dispatch]);
+
   return (
     <div className={cl.promocodeManagement}>
       <div>
